refactor: group type exports
