+++ src/types/index.ts
@@ -1,7 +1,8 @@
-export * from './lib';
-export * from './signer';
-export * from './contract';
 export * from './account';
+export * from './contract';
+export * from './lib';
 export * from './provider';
+export * from './signer';
+export * from './typedData';
 export * from './api/sequencer';
 export * from './api/rpc';
