+++ src/types/provider/configuration.ts
@@ -1,0 +1,32 @@
+import { NetworkName, StarknetChainId } from '../../constants';
+import { BlockIdentifier } from '../lib';
+
+export interface ProviderOptions {
+  sequencer?: SequencerProviderOptions;
+  rpc?: RpcProviderOptions;
+}
+
+export type RpcProviderOptions = {
+  nodeUrl: string;
+  retries?: number;
+  headers?: object;
+  blockIdentifier?: BlockIdentifier;
+  chainId?: StarknetChainId;
+};
+
+export type SequencerHttpMethod = 'POST' | 'GET';
+
+export type SequencerProviderOptions = {
+  headers?: Record<string, string>;
+  blockIdentifier?: BlockIdentifier;
+  chainId?: StarknetChainId;
+} & (
+  | {
+      network: NetworkName | StarknetChainId;
+    }
+  | {
+      baseUrl: string;
+      feederGatewayUrl?: string;
+      gatewayUrl?: string;
+    }
+);
